[Rebuild] Feed NavigationBar item 객체들 외부에 정의한 constant 내부에 추가 및 코드 명료화
